+++ src/fields/formfields/Invisible.php
@@ -2,12 +2,19 @@
 
 namespace barrelstrength\sproutforms\fields\formfields;
 
+use barrelstrength\sproutforms\SproutForms;
 use Craft;
 use craft\base\ElementInterface;
 use craft\helpers\Template as TemplateHelper;
 use craft\base\PreviewableFieldInterface;
 use barrelstrength\sproutforms\base\FormField;
 
+/**
+ *
+ * @property string $svgIconPath
+ * @property mixed  $settingsHtml
+ * @property mixed  $exampleInputHtml
+ */
 class Invisible extends FormField implements PreviewableFieldInterface
 {
     /**
@@ -97,10 +104,11 @@ class Invisible extends FormField implements PreviewableFieldInterface
 
     /**
      * @inheritdoc
+     * @throws \craft\errors\MissingComponentException
      */
     public function getFrontEndInputHtml($value, array $renderingOptions = null): string
     {
-        Craft::$app->getSession()->set($this->handle, $this->value);
+        $this->preProcessInvisibleValue();
 
         $html = '<input type="hidden" name="'.$this->handle.'">';
 
@@ -109,13 +117,20 @@ class Invisible extends FormField implements PreviewableFieldInterface
 
     /**
      * @inheritdoc
+     *
+     * @param                       $value
+     * @param ElementInterface|null $element
+     *
+     * @return string
+     * @throws \Throwable
+     * @throws \craft\errors\MissingComponentException
+     * @throws \yii\base\Exception
      */
     public function normalizeValue($value, ElementInterface $element = null)
     {
-        $session = Craft::$app->getSession()->get($this->handle);
-        $sessionValue = $session ?: '';
+        $value = Craft::$app->getSession()->get($this->handle) ?? '';
 
-        return Craft::$app->view->renderObjectTemplate($sessionValue, parent::getFieldVariables());
+        return $value;
     }
 
     /**
@@ -130,5 +145,25 @@ class Invisible extends FormField implements PreviewableFieldInterface
         }
 
         return $hiddenValue;
+    }
+
+    /**
+     * @return string
+     * @throws \Throwable
+     */
+    private function preProcessInvisibleValue()
+    {
+        $value = '';
+
+        if ($this->value) {
+            try {
+                $value = Craft::$app->view->renderObjectTemplate($this->value, parent::getFieldVariables());
+                Craft::$app->getSession()->set($this->handle, $value);
+            } catch (\Exception $e) {
+                SproutForms::error($e->getMessage());
+            }
+        }
+
+        return $value;
     }
 }